// src/expandableCalendar/commons.d.ts
export declare const todayString = "today";
export declare enum UpdateSources {
    CALENDAR_INIT = "calendarInit",
    PROP_UPDATE = "propUpdate",
    TODAY_PRESS = "todayPress",
    DAY_PRESS = "dayPress",
    ARROW_PRESS = "arrowPress",
    WEEK_ARROW_PRESS = "weekArrowPress",
    LIST_DRAG = "listDrag",
    PAGE_SCROLL = "pageScroll",
    WEEK_SCROLL = "weekScroll"
}
export declare enum CalendarNavigationTypes {
    AGENDA_SCROLL = "listDrag",
    MONTH_SCROLL = "pageScroll",
    WEEK_SCROLL = "weekScroll",
    MONTH_ARROWS = "arrowPress",
    WEEK_ARROWS = "weekArrowPress",
    TODAY_PRESS = "todayPress"
}